Implement Base62 encoder/decoder
The encoding is very simple and has limited set of input values
to only accept positive Long. This is due to the fact that it is
supposed to be used only to encode/decode CRC32 algorithm output,
which is guaranteed to be non-negative Long. Besides, the current
implementation would return same values for a number and its
additive inverse.
For the 2 reasons above it was decided to limit the input space.