Change values of JwtPermissions to contain prefix with service name
This is a practise I have seen being used by Tableau.
It allows for clear separation of permissions when a service
is used as a third party, which is also the case for api-key-steward.
It protects from users or admins being granted access to
the third party by a coincidence of the same permission values.