Add SQL file with instructions to add trigram indexes for permission.name column
There are SELECT queries on permission.name column which use 'ILIKE' keyword.
LIKE and ILIKE searches are known to be very slow, but their performance can be
improved with the use of these trigram indexes (at the cost of inserts and space).

Creating these indexes is not part of DB migrations, because in my opinion
it should be optional to use them. The benefit of using these indexes may not be
noticeable for the first couple thousand rows.
So it may not be necessary for smaller users.